Refactor commands; add commands to autosuggestions
