+++ src/renderer/components/battle/BattleChat.vue
@@ -28,32 +28,15 @@
 import { Icon } from "@iconify/vue";
 import eyeIcon from "@iconify-icons/mdi/eye";
 import type { Ref } from "vue";
-import { ref } from "vue";
+import { onMounted, onUnmounted, reactive, ref, watch } from "vue";
 
+import { Command, getAutoSuggestions, serverCommandList } from "@/api/commands";
 import BattleMessage from "@/components/battle/BattleMessage.vue";
 import AutoSuggest from "@/components/controls/AutoSuggest.vue";
 import AutoSuggestionOption from "@/utils/auto-suggestion-option";
 
-const optionList: Ref<AutoSuggestionOption[]> = ref([
-    {
-        suggestion: "/Hello (there)",
-        description: "Says hello",
-        replaceSuggestion: "$hello",
-    },
-    {
-        suggestion: "/hello 2",
-        description: "Says hell",
-    },
-    {
-        suggestion: "hel",
-        description: "Says hel",
-    },
-    {
-        suggestion: "/Hello (you)",
-        description: "Says hello you",
-        replaceSuggestion: "!hello",
-    },
-]);
+const optionList: Ref<AutoSuggestionOption[]> = ref([]);
+const commandList = reactive<Command[]>(structuredClone(serverCommandList));
 const messages = api.session.battleMessages;
 const myMessage = ref("");
 const showHiddenMessages = ref(false);
@@ -75,6 +58,34 @@ function sendMessage() {
 
     myMessage.value = "";
 }
+
+watch(
+    commandList,
+    (newCommandList) => {
+        optionList.value = getAutoSuggestions(newCommandList);
+    },
+    { deep: true }
+);
+
+const commandListener = api.comms.onResponse("s.communication.received_direct_message").add(async (data) => {
+    const { message } = data;
+    // Check if the message is a command
+    if (!message.startsWith("!") && !message.startsWith("$")) return;
+    const cmd = message.split("-")[0].split(" ")[0];
+    const cmdDescription = message.slice(cmd.length + 1).replace("-", " ");
+    cmdDescription && !cmdDescription.includes("*") && commandList.push({ cmd, cmdDescription });
+});
+
+onMounted(() => {
+    api.comms.request("c.communication.send_direct_message", {
+        recipient_id: 3137,
+        message: `!helpall`,
+    });
+});
+
+onUnmounted(() => {
+    commandListener.destroy();
+});
 </script>
 
 <style lang="scss" scoped>
